+++ xrpl4j-core/src/main/java/org/xrpl/xrpl4j/model/jackson/modules/SignatureSerializer.java
@@ -1,15 +1,16 @@
-package org.xrpl.xrpl4j.crypto.core.signing;
+package org.xrpl.xrpl4j.model.jackson.modules;
 
 import com.fasterxml.jackson.core.JsonGenerator;
 import com.fasterxml.jackson.databind.JsonSerializer;
 import com.fasterxml.jackson.databind.SerializerProvider;
+import org.xrpl.xrpl4j.crypto.core.signing.Signature;
 
 import java.io.IOException;
 
 /**
  * Custom Jackson serializer for {@link Signature}es.
  */
-class SignatureSerializer extends JsonSerializer<Signature> {
+public class SignatureSerializer extends JsonSerializer<Signature> {
   @Override
   public void serialize(Signature signature, JsonGenerator gen, SerializerProvider provider) throws IOException {
     gen.writeString(signature.base16Value());
