+++ xrpl4j-core/src/main/java/org/xrpl/xrpl4j/model/jackson/modules/Xrpl4jModule.java
@@ -23,6 +23,7 @@ package org.xrpl.xrpl4j.model.jackson.modules;
 import com.fasterxml.jackson.core.Version;
 import com.fasterxml.jackson.databind.module.SimpleModule;
 import org.xrpl.xrpl4j.codec.addresses.UnsignedByteArray;
+import org.xrpl.xrpl4j.crypto.core.signing.ImmutableSignature;
 import org.xrpl.xrpl4j.model.client.accounts.AccountTransactionsRequestParams;
 import org.xrpl.xrpl4j.model.client.accounts.AccountTransactionsTransaction;
 import org.xrpl.xrpl4j.model.client.accounts.GatewayBalancesAssets;
@@ -111,5 +112,8 @@ public class Xrpl4jModule extends SimpleModule {
     
     addSerializer(UnsignedByteArray.class, new UnsignedByteArraySerializer());
     addDeserializer(UnsignedByteArray.class, new UnsignedByteArrayDeserializer());
+
+    addSerializer(ImmutableSignature.class, new SignatureSerializer());
+    addDeserializer(ImmutableSignature.class, new SignatureDeserializer());
   }
 }
